+++ main/src/main/java/ru/practicum/main/user/dao/UserRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.domain.Pageable;
 
 import java.util.List;
 
-public interface UserRepository extends JpaRepository<User, Long> {
-    List<User> findByIdIn(Long[] ids, Pageable page);
+public interface UserRepository extends JpaRepository<User, Integer> {
+    List<User> findByIdIn(Integer[] ids, Pageable page);
 }
